--- app/src/preview/componentPreview/constants.ts
@@ -1,3 +1,0 @@
-export const PlayWrightInterfaceName = '_uui_playwright_interface';
-
-export const SCREENSHOT_WIDTH_LIMIT = 1280;
