+++ app/src/preview/previewContent/previewLayout.module.scss
@@ -4,7 +4,6 @@
 }
 
 .root {
-    /* test automation screenshots rely on this width to be exactly equal EXACTLY 1280px. do not change. */
     flex-direction: column;
     align-items: flex-start;
     background-color: transparent;
