+++ app/src/preview/previewPage.tsx
@@ -1,12 +1,15 @@
+import React, { useCallback, useMemo } from 'react';
 import { Page } from '../common';
-import React, { useCallback, useEffect, useMemo } from 'react';
-import { ComponentPreview } from './componentPreview/componentPreview';
-import { PreviewQueryHelpers, usePreviewParams } from './previewQueryHelpers';
-import { TComponentPreviewParams } from './componentPreview/types';
-import { PlayWrightInterfaceName } from './componentPreview/constants';
+import { PreviewContent } from './previewContent/previewContent';
+import { TPreviewContentParams } from './types';
+import { usePlayWrightInterface } from './hooks/usePlayWrightInterface';
+import { usePreviewPageBg } from './hooks/usePreviewPageBg';
+import { usePreviewParams } from './hooks/usePreviewParams';
+import { svc } from '../services';
+import { TTheme } from '../common/docs/docsConstants';
+import { formatPreviewIdToString } from './utils/previewLinkUtils';
 
 import css from './previewPage.module.scss';
-import { useLayoutEffectSafeForSsr } from '@epam/uui-core';
 
 export function PreviewPage() {
     const params = usePreviewParams();
@@ -15,7 +18,7 @@ export function PreviewPage() {
     const componentId = params.componentId;
     const previewId = params.previewId;
 
-    const currentParams: TComponentPreviewParams = useMemo(() => {
+    const currentParams: TPreviewContentParams = useMemo(() => {
         return {
             theme,
             isSkin,
@@ -29,45 +32,25 @@ export function PreviewPage() {
         previewId,
     ]);
 
-    const handleNavPreview = useCallback((newParams: TComponentPreviewParams) => {
-        PreviewQueryHelpers.setParams({
-            componentId: newParams.componentId,
-            previewId: newParams.previewId,
-            theme: newParams.theme,
-            isSkin: newParams.isSkin,
+    const handleNavPreview = useCallback((newParams: TPreviewContentParams) => {
+        svc.uuiRouter.redirect({
+            pathname: '/preview',
+            query: {
+                theme: newParams.theme || TTheme.promo,
+                isSkin: newParams.isSkin ?? true,
+                componentId: newParams.componentId,
+                previewId: formatPreviewIdToString(newParams.previewId),
+            },
         });
     }, []);
 
     usePlayWrightInterface(handleNavPreview);
-
-    useLayoutEffectSafeForSsr(() => {
-        const style = document.body.style;
-        const prev = style.backgroundColor;
-        style.backgroundColor = 'var(--uui-surface-main)';
-        return () => {
-            style.backgroundColor = prev;
-        };
-    }, []);
+    usePreviewPageBg();
 
     const key = `${theme}_${isSkin}_${componentId}_${previewId}`;
     return (
         <Page renderHeader={ () => null } rootCx={ css.root }>
-            <ComponentPreview
-                key={ key }
-                params={ currentParams }
-                onParamsChange={ handleNavPreview }
-            />
+            <PreviewContent key={ key } params={ currentParams } />
         </Page>
     );
-}
-
-function usePlayWrightInterface(setter: (newParams: TComponentPreviewParams) => void) {
-    useEffect(() => {
-        (window as any)[PlayWrightInterfaceName] = (_params: string) => {
-            setter(JSON.parse(_params) as TComponentPreviewParams);
-        };
-        return () => {
-            delete (window as any)[PlayWrightInterfaceName];
-        };
-    }, [setter]);
 }
